requirements.txt
asgiref==3.7.2
Django==4.2.3
django-fontawesome-5==1.0.18
django-froala-editor==4.1.0
django-phonenumber-field==7.1.0
install==1.3.5
phonenumbers==8.13.16
Pillow==10.0.0
sqlparse==0.4.4
tzdata==2023.3
Unidecode==1.3.6
